% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example1_PS6}
\alias{example1_PS6}
\title{Parameters required for Example 1}
\usage{
example1_PS6(Nsim)
}
\arguments{
\item{Nsim}{Number of simulations}
}
\description{
Parameters required for Example 1
}

