% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{example2_PS6}
\alias{example2_PS6}
\title{Parameters required for Example 2}
\usage{
example2_PS6(Nsim, n)
}
\arguments{
\item{Nsim}{Number of simulations}
}
\description{
Parameters required for Example 2
}

