% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{sample_dataARMA}
\alias{sample_dataARMA}
\title{Sample Data from an ARMA model}
\usage{
sample_dataARMA(Phi0, Theta0, Sigma0, p0, q0, m, n, ...)
}
\arguments{
\item{...}{}
}
\description{
Sample Data from an ARMA model
}

