% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_criteria.R
\name{select_model}
\alias{select_model}
\title{Compute frequency of selected criteria}
\usage{
select_model(results, NumLags)
}
\arguments{
\item{results}{(Nsim x NumLags x 4) array with computed criteria}

\item{NumLags}{Maximum number of lags considered for evaluation}
}
\value{
4 x NumLags matrix with AIC,AICc, AICcBD and SIC criteria,
on each row, respectively.
}
\description{
Compute frequency of selected criteria
}

